+++ server/src/main/java/com/main/project/restaurant/dto/RestaurantPatchResponseDto.java
@@ -19,8 +19,8 @@ public class RestaurantPatchResponseDto {
     private double aveTaste;                // 맛 별점 평균
     private double aveFacility;             // 시설 별점 평균
     private double avePrice;                // 가격 별점 평균
-    private int mapx;
-    private int mapy;
+    private double mapx;
+    private double mapy;
     private long locationId;
     String foodTypeName;
 }
